Added "Pixie Seal" example
Wrote a new "toy problem" for illustrating the issue. 

It's a population model for fictional pixie seals (concept art may be added later) that uses latent multinomial likelihoods on two occasions: 
1) Initializing population model ("partitioning" initial population size into 4 stages using fixed, known proportions)
2) Projecting stage-structured population dynamics (survival and stage transitions over time)

The "dmulti_dmulti" version used multinomial likelihoods for both processes above. 
The three other versions swap out either or both with alternative - but less elegant - (sequential) binomial likelihoods to illustrate that it is indeed the multinomial likelihoods that result in the "fixing" of a range of nodes.